# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/permissions": {
      "post": {
        "operationId": "PermissionsController_create",
        "x-required-roles": [
          "admin"
        ],
        "summary": "[ADMIN APP] Create a new permission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin App - Permissions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Admin App",
        "x-audience-label": "Admin App"
      },
      "get": {
        "operationId": "PermissionsController_findAll",
        "summary": "[ADMIN APP] Retrieve all permissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin App - Permissions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Admin App",
        "x-audience-label": "Admin App"
      }
    },
    "/api/v1/permissions/resources": {
      "get": {
        "operationId": "PermissionsController_getAllResources",
        "summary": "[ADMIN APP] Retrieve all resources",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin App - Permissions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Admin App",
        "x-audience-label": "Admin App"
      }
    },
    "/api/v1/permissions/actions": {
      "get": {
        "operationId": "PermissionsController_getAllActions",
        "summary": "[ADMIN APP] Retrieve all actions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin App - Permissions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Admin App",
        "x-audience-label": "Admin App"
      }
    },
    "/api/v1/permissions/by-resource": {
      "get": {
        "operationId": "PermissionsController_findByResource",
        "summary": "[ADMIN APP] Retrieve all permissions by resource",
        "parameters": [
          {
            "name": "resource",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin App - Permissions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Admin App",
        "x-audience-label": "Admin App"
      }
    },
    "/api/v1/permissions/{id}": {
      "get": {
        "operationId": "PermissionsController_findOne",
        "summary": "[ADMIN APP] Retrieve a specific permission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin App - Permissions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Admin App",
        "x-audience-label": "Admin App"
      },
      "patch": {
        "operationId": "PermissionsController_update",
        "x-required-roles": [
          "admin"
        ],
        "summary": "[ADMIN APP] Update a specific permission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin App - Permissions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Admin App",
        "x-audience-label": "Admin App"
      },
      "delete": {
        "operationId": "PermissionsController_remove",
        "x-required-roles": [
          "admin"
        ],
        "summary": "[ADMIN APP] Use Guards",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin App - Permissions"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "description": "Audience: Admin App",
        "x-audience-label": "Admin App"
      }
    },
...
  },
  "info": {
    "title": "La Favola Pizza Restaurant API",
    "description": "Routes are grouped by audience. Customer app developers should use sections prefixed \"Customer App\" and \"Audience: Public / Customer App\". Administrative and staff operations declare their required roles in their Audience tag and x-required-roles metadata.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the permission",
            "example": "Create User",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the permission",
            "example": "Allows creating a new user in the system",
            "maxLength": 255
          },
          "resource": {
            "type": "string",
            "description": "Resource the permission applies to",
            "example": "user",
            "maxLength": 50
          },
          "action": {
            "type": "string",
            "description": "Action allowed by this permission",
            "example": "create",
            "maxLength": 50
          }
        },
        "required": [
          "name",
          "resource",
          "action"
        ]
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the permission",
            "example": "Create User",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the permission",
            "example": "Allows creating a new user in the system",
            "maxLength": 255
          },
          "resource": {
            "type": "string",
            "description": "Resource the permission applies to",
            "example": "user",
            "maxLength": 50
          },
          "action": {
            "type": "string",
            "description": "Action allowed by this permission",
            "example": "create",
            "maxLength": 50
          }
        }
      },
...
    }
  },
  "security": [
    {
      "JWT-auth": []
    }
  ]
}